requirements.txt
pygame>=2.1.0
boto3>=1.26.0
pyjwt>=2.6.0
pyyaml>=6.0
